refactor: move StatsEvent to stats
